refactor(hooks): finalize migration, remove legacy services and rename useApiTest to useApi
