+++ src/app/(pages)/[username]/notes/page.tsx
@@ -5,7 +5,7 @@ import { Element } from "./elements";
 import { IconEyeOff, IconLock, IconNotesOff } from "@tabler/icons-react";
 import { Section } from "../components/Section";
 import { useParams, useSearchParams } from "next/navigation";
-import { useUser, useTags, useApiTest } from "@/data/hooks";
+import { useUser, useTags, useApi } from "@/data/hooks";
 
 const Page = () => {
 
@@ -19,7 +19,7 @@ const Page = () => {
     const accessToken = token ? token.access_token : null;
     const isCurrentUser = user ? user.username === username : false;
 
-    const { noteQueries: { useFindUserTags, useFindUserNotes } } = useApiTest();
+    const { noteQueries: { useFindUserTags, useFindUserNotes } } = useApi();
 
     const { data: tagsResponse, isLoading: tagsLoading } = useFindUserTags(
         accessToken,
