+++ src/app/(pages)/settings/(pages)/account/(pages)/delete/elements/OAuthUserTitle.tsx
@@ -1,6 +1,6 @@
 import { clsx } from "clsx";
 import { IconSend } from "@tabler/icons-react";
-import { useApiTest, useUser } from "@/data/hooks";
+import { useApi, useUser } from "@/data/hooks";
 import { useState } from "react";
 
 export const OAuthUserTitle = (props: React.HTMLAttributes<HTMLElement>) => {
@@ -8,7 +8,7 @@ export const OAuthUserTitle = (props: React.HTMLAttributes<HTMLElement>) => {
     const {
         authService: { sendSecretKeyRequest },
         withProgress
-    } = useApiTest();
+    } = useApi();
     const { user } = useUser();
 
     const [isPending, setIsPending] = useState<boolean>(false);
