+++ src/app/(pages)/settings/(pages)/account/(pages)/email/page.tsx
@@ -3,7 +3,7 @@
 import { clsx } from "clsx";
 import { Header } from "../../../Header";
 import { IconSend } from "@tabler/icons-react";
-import { useApiTest, useUser } from "@/data/hooks";
+import { useApi, useUser } from "@/data/hooks";
 import { useState } from "react";
 
 const Page = () => {
@@ -11,7 +11,7 @@ const Page = () => {
     const {
         authService: { sendEmailChangeRequest },
         withProgress
-    } = useApiTest();
+    } = useApi();
     const { user } = useUser();
 
     const [isPending, setIsPending] = useState<boolean>(false);
