+++ src/components/forms/auth/recover/index.tsx
@@ -2,7 +2,7 @@ import { Element } from "./elements";
 import { FormProvider, useForm } from "react-hook-form";
 import { handleFieldErrors, RecoverFormData, recoverFormSchema } from "@/core";
 import { IconMail } from "@tabler/icons-react";
-import { useApiTest } from "@/data/hooks";
+import { useApi } from "@/data/hooks";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { zodResolver } from "@hookform/resolvers/zod";
@@ -12,7 +12,7 @@ export const Form = (props: React.FormHTMLAttributes<HTMLFormElement>) => {
     const {
         authService: { sendPasswordChangeRequest },
         withProgress
-    } = useApiTest();
+    } = useApi();
 
     const recoverForm = useForm<RecoverFormData>({
         resolver: zodResolver(recoverFormSchema)
