+++ src/components/forms/auth/signin/index.tsx
@@ -3,7 +3,7 @@ import { Element } from "./elements";
 import { FormProvider, useForm } from "react-hook-form";
 import { handleFieldErrors, LoginFormData, loginFormSchema, Token, User } from "@/core";
 import { IconAt } from "@tabler/icons-react";
-import { useApiTest, useServices, useStore, useUser } from "@/data/hooks";
+import { useApi, useStore, useUser } from "@/data/hooks";
 import { useCallback, useEffect, useRef, useState } from "react";
 import { useGoogleLogin } from "@react-oauth/google";
 import { useRouter, useSearchParams } from "next/navigation";
@@ -15,7 +15,7 @@ export const Form = (props: React.FormHTMLAttributes<HTMLFormElement>) => {
     const {
         authService: { loginUserByDefault, loginUserByGoogle, loginUserByGitHub },
         withProgress
-    } = useApiTest();
+    } = useApi();
 
     const { setStore } = useStore();
     const { setUser } = useUser();
