+++ src/components/forms/auth/signup/index.tsx
@@ -2,7 +2,7 @@ import { CreateUserFormData, createUserFormSchema, handleFieldErrors } from "@/c
 import { Element } from "./elements";
 import { FormProvider, useForm } from "react-hook-form";
 import { IconAt, IconMail, IconSignature } from "@tabler/icons-react";
-import { useApiTest } from "@/data/hooks";
+import { useApi } from "@/data/hooks";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { zodResolver } from "@hookform/resolvers/zod";
@@ -12,7 +12,7 @@ export const Form = (props: React.FormHTMLAttributes<HTMLFormElement>) => {
     const {
         userService: { createUser },
         withProgress
-    } = useApiTest();
+    } = useApi();
 
     const createUserForm = useForm<CreateUserFormData>({
         resolver: zodResolver(createUserFormSchema)
