+++ src/components/forms/reply/update/index.tsx
@@ -5,7 +5,7 @@ import { FormProvider, useForm } from "react-hook-form";
 import { IconEdit, IconX } from "@tabler/icons-react";
 import { InfiniteData, useQueryClient } from '@tanstack/react-query';
 import { Menu, MenuButton, MenuItem } from "@/components/menu";
-import { useApiTest } from "@/data/hooks";
+import { useApi } from "@/data/hooks";
 import { useRef, useState, useTransition } from "react";
 import { zodResolver } from "@hookform/resolvers/zod";
 
@@ -24,7 +24,7 @@ type RepliesCache = InfiniteData<Page<Reply>>;
 
 export const Form = ({ token, user, note, comment, reply, setReplies, setRepliesCount, setIsReplying, ...rest }: FormProps) => {
 
-    const { replyService: { editReply, deleteReply } } = useApiTest();
+    const { replyService: { editReply, deleteReply } } = useApi();
     const qc = useQueryClient();
 
     const editReplyForm = useForm<CreateReplyFormData>({
