+++ src/components/forms/user/update-password/index.tsx
@@ -1,7 +1,7 @@
 import { Element } from "./elements";
 import { FormProvider, useForm, useWatch } from "react-hook-form";
 import { handleFieldErrors, handleInvalidTokenFieldError, PasswordUpdateFormData, passwordUpdateFormSchema } from "@/core";
-import { useApiTest, useUser } from "@/data/hooks";
+import { useApi, useUser } from "@/data/hooks";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { zodResolver } from "@hookform/resolvers/zod";
@@ -11,7 +11,7 @@ export const Form = ({ token }: { token: string }) => {
     const {
         userService: { updateUserPassword },
         withProgress
-    } = useApiTest();
+    } = useApi();
     const { clearUser } = useUser();
 
     const passwordUpdateForm = useForm<PasswordUpdateFormData>({
