+++ src/data/hooks/index.ts
@@ -1,6 +1,4 @@
-export * from './useAPI';
-export * from './useApiTest';
-export * from './useServices';
+export * from './useApi';
 export * from './usePreferences';
 export * from './useSessionPreferences';
 export * from './useStore';
