+++ src/data/hooks/useAPI.ts
@@ -1,94 +1,49 @@
-import { useCallback } from "react";
-import { useProgress } from "./useProgress";
-import { useStore } from "./useStore";
-import { UUID } from "crypto";
+import { createApiClient } from '@/api';
+import { createAuthService, createUserService, createUserQuery, createNoteService, createNoteQuery, createFlameService, createFlameQuery, createFeedService, createFeedQuery, createReplyService, createReplyQuery, createCommentService, createCommentQuery, createSponsorshipService, createHealthService } from '@/services';
+import { useMemo } from 'react';
+import { useProgress } from './useProgress';
+import { useStore } from './useStore';
+import { useUser } from './useUser';
 
-interface HttpOptions {
-    useProgress?: boolean;
-    useRefreshToken?: string;
-    useToken?: string | null;
-}
-
-const baseUrl = process.env.NEXT_PUBLIC_API_URL;
-
-const createHeaders = (
-    useDeviceId: UUID | string,
-    useRefreshToken?: string,
-    useToken?: string | null
-): HeadersInit => {
-    const headers: HeadersInit = { 'X-Device-Id': useDeviceId, 'Content-Type': 'application/json' };
-    if (useRefreshToken) headers['X-Refresh-Token'] = `${useRefreshToken}`;
-    if (useToken) headers['Authorization'] = `Bearer ${useToken}`;
-    return headers;
-}
-
-const handleResponse = async (response: Response) => {
-
-    if (response.status === 204) return null;
-
-    const text = await response.text();
-    let data;
-    try {
-        data = text ? JSON.parse(text) : null;
-    } catch {
-        data = text;
-    }
-
-    if (response.ok) return data;
-    throw { response, data };
-
-}
-
-export const useAPI = () => {
+export const useApi = () => {
 
     const { store: { device } } = useStore();
     const { setOnProgress } = useProgress();
+    const { token, updateToken } = useUser();
 
-    const request = useCallback(async (method: string, endpoint: string, body?: any, options?: HttpOptions) => {
+    const api = useMemo(() => createApiClient({
+        deviceId: device,
+        token: token ? token.access_token : null,
+    }), [device, token])
 
-        const { useProgress: showProgress, useRefreshToken, useToken } = options || {};
+    const services = useMemo(() => ({
+        healthService: createHealthService(api),
+        authService: createAuthService(api, updateToken),
+        userService: createUserService(api, updateToken),
+        sponsorshipService: createSponsorshipService(api, updateToken),
+        feedService: createFeedService(api, updateToken),
+        noteService: createNoteService(api, updateToken),
+        flameService: createFlameService(api, updateToken),
+        commentService: createCommentService(api, updateToken),
+        replyService: createReplyService(api, updateToken),
+    }), [api, updateToken])
 
-        const uri = `${baseUrl}${endpoint.startsWith('/') ? endpoint : `/${endpoint}`}`;
+    const queries = useMemo(() => ({
+        userQueries: createUserQuery(services.userService),
+        feedQueries: createFeedQuery(services.feedService),
+        noteQueries: createNoteQuery(services.noteService),
+        flameQueries: createFlameQuery(services.flameService),
+        commentQueries: createCommentQuery(services.commentService),
+        replyQueries: createReplyQuery(services.replyService),
+    }), [services])
 
-        const headers = createHeaders(device, useRefreshToken, useToken);
+    const withProgress = useMemo(() =>
+        async <T>(fn: () => Promise<T>): Promise<T> => {
+            setOnProgress(true);
+            try { return await fn(); }
+            finally { setOnProgress(false); }
+        }, [setOnProgress])
 
-        if (showProgress) setOnProgress(true);
-
-        const config: RequestInit = {
-            method,
-            headers,
-            body: body ? JSON.stringify(body) : undefined,
-        }
-
-        try {
-            const response = await fetch(uri, config);
-            return await handleResponse(response);
-        } finally {
-            if (showProgress) setOnProgress(false);
-        }
-
-    }, [device, setOnProgress]);
-
-    const httpPost = useCallback((endpoint: string, body: any, options?: HttpOptions) => {
-        return request('POST', endpoint, body, options);
-    }, [request])
-
-    const httpPut = useCallback((endpoint: string, body: any, options?: HttpOptions) => {
-        return request('PUT', endpoint, body, options);
-    }, [request])
-
-    const httpPatch = useCallback((endpoint: string, body: any, options?: HttpOptions) => {
-        return request('PATCH', endpoint, body, options);
-    }, [request])
-
-    const httpGet = useCallback((endpoint: string, options?: HttpOptions) => {
-        return request('GET', endpoint, undefined, options);
-    }, [request])
-
-    const httpDelete = useCallback((endpoint: string, body: any, options?: HttpOptions) => {
-        return request('DELETE', endpoint, body, options);
-    }, [request])
-
-    return { httpPost, httpPut, httpPatch, httpGet, httpDelete };
+    return { ...services, ...queries, withProgress };
 
 }
