--- src/data/hooks/useServices.ts
@@ -1,3 +1,0 @@
-export const useServices = () => {
-    return {}
-}
